refactor(frontend): server-only Data Access Layer for auth flow
Centralizes the OpenAI-key session lifecycle (verify → seal+store →
read+unseal → clear) in `lib/data/auth.ts`. Server Actions and the
`/api/chat` route handler now delegate to one source of truth instead
of each duplicating cookies + AES-256-GCM + key-shape validation at
the call site.

Pattern follows the Next.js Data Security guide's Data Access Layer
recommendation. Module is server-only by virtue of `next/headers` —
that import throws on the client, so accidental client-side use fails
fast at runtime without needing the `server-only` package directive
(which conflicts with the test-fixture import of `seal` from
`session-crypto` in `tests/e2e/session-persistence.spec.ts`).

Also pins `retries: 2` in `playwright.config.ts`. Browser-driven
sessionStorage-write → reload → React-mount → restored-DOM
assertions have inherent cross-process timing flakiness (slow GC
pause, kernel scheduling) that surfaces as occasional first-attempt
failures. Two retries cover real flakes without masking genuine
regressions — a true product bug fails all three runs.
